refactor(components): re-structured components folder.
- moved all slider components in 'Sliders' folder.
- moved 'SearchInput.jsx' to 'Inputs' folder.
- moved 'Player.jsx' to 'Shared' folder.
- fixed all import statements.
