+++ src/components/shared/HamburgerMenu/HamburgerMenu.jsx
@@ -1,7 +1,7 @@
 import { cloneElement, memo } from 'react';
 import { createPortal } from 'react-dom';
-import useHamburgerMenu from '../../hooks/useHamburgerMenu';
-import Avatar from '../Avatar/Avatar';
+import useHamburgerMenu from '../../../hooks/useHamburgerMenu';
+import Avatar from '../../Avatar/Avatar';
 import {
   Heart,
   Home2,
