+++ src/pages/Browse/Browse.jsx
@@ -1,11 +1,11 @@
 import SidebarPlaylist from '../../components/SidebarPlaylist/SidebarPlaylist';
 import SectionTitle from '../../components/SectionHeader/SectionHeader';
-import TracksSlider from '../../components/TracksSlider/TracksSlider';
-import PlaylistsSlider from '../../components/PlaylistsSlider/PlaylistsSlider';
-import DiscoverPlaylistsSlider from '../../components/DiscoverPlaylistsSlider/DiscoverPlaylistsSlider';
-import ArtistsSlider from '../../components/ArtistsSlider/ArtistsSlider';
-import GenresSlider from '../../components/GenresSlider/GenresSlider';
-import AlbumsSlider from '../../components/AlbumsSlider/AlbumsSlider';
+import TracksSlider from '../../components/Sliders/TracksSlider/TracksSlider';
+import PlaylistsSlider from '../../components/Sliders/PlaylistsSlider/PlaylistsSlider';
+import DiscoverPlaylistsSlider from '../../components/Sliders/DiscoverPlaylistsSlider/DiscoverPlaylistsSlider';
+import ArtistsSlider from '../../components/Sliders/ArtistsSlider/ArtistsSlider';
+import GenresSlider from '../../components/Sliders/GenresSlider/GenresSlider';
+import AlbumsSlider from '../../components/Sliders/AlbumsSlider/AlbumsSlider';
 import PlaylistCard from '../../components/MusicCards/PlaylistCard/PlaylistCard';
 import MainButton from '../../components/Buttons/MainButton/MainButton';
 import useMediaQuery from '../../hooks/useMediaQuery';
